requirements.txt
aiosmtplib==2.0.2
annotated-types==0.7.0
bcrypt==4.2.0
cryptography==43.0.0
email_validator==2.2.0
fastapi==0.112.0
fastapi-mail==1.4.1
httpcore==1.0.5
httpx==0.27.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
passlib==1.7.4
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
PyJWT==2.9.0
pytest==8.3.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
SQLAlchemy==2.0.31
starlette==0.37.2
typing_extensions==4.12.2
uvicorn==0.30.4
sentry-sdk[fastapi]
asyncpg
supabase
aiosqlite
pytest-asyncio


